apps/api/pyproject.toml
[project]
name = "navigator-api"
version = "0.1.0"
description = ""
authors = [
    {name = "ishish.io",email = "ts@ishish.io"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi (>=0.128.0,<0.129.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "pyjwt (>=2.11.0,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "cryptography (>=46.0.4,<47.0.0)",
    "customer-accounts @ file:///Users/tsalacinski/Projects/2026-02-08-navigator-2/ai-navigator/packages/modules/customer_accounts",
    "navigator-tests @ file:///Users/tsalacinski/Projects/2026-02-08-navigator-2/ai-navigator/packages/modules/navigator_tests",
    "navigator-platform @ file:///Users/tsalacinski/Projects/2026-02-08-navigator-2/ai-navigator/packages/platform",
]
packages = [{include = "navigator_api", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
customer-accounts = {develop = true}
navigator-tests = {develop = true}
navigator-platform = {develop = true}

[tool.poetry.group.dev.dependencies]
pytest = "^9.0.2"
httpx = "^0.28.1"

